Implemented Survival analysis as a classification problem
we explore a method for treating survival analysis as a
classi cation problem. The method uses a \stacking" idea that collects the
features and outcomes of the survival data in a large data frame, and then
treats it as a classi cation problem. In this framework, various statistical
learning algorithms (including logistic regression, random forests, gradi-
ent boosting machines and neural networks) can be applied to estimate
the parameters and make predictions.